Add AddTaskCommand to handle when tasks are added
