+++ src/main/java/duke/commands/AddEventCommand.java
@@ -6,23 +6,10 @@ import duke.utils.TaskList;
 import duke.tasks.Event;
 import duke.utils.Ui;
 
-public class AddEventCommand extends Command {
+public class AddEventCommand extends AddTaskCommand {
 
     private String[] keyWords;
-    public AddEventCommand(String[] keywords) {
-        this.keyWords = keywords;
-    }
-
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws InvalidTimeStampException {
-        Event event = new Event(this.keyWords[0], this.keyWords[1]);
-        tasks.addTask(event);
-        String[] messages = new String[] {
-                "Got it. I've added this task:",
-                "    " + event.toString(),
-                String.format("Now you have %d tasks in the list.", tasks.getSize())
-        };
-        ui.printOut(messages);
-        storage.save(tasks);
+    public AddEventCommand(String[] keywords) throws InvalidTimeStampException {
+        super(new Event(keywords[0], keywords[1]));
     }
 }
